refactor: remove ibc/testing/simapp
